добавил проверку для институтов

+++ src/main/java/io/proj3ct/ReturnBot1/keybords/LogicForDeleteDepartment.java
@@ -4,8 +4,13 @@ import io.proj3ct.ReturnBot1.baseClasses.EnvironmentService;
 import io.proj3ct.ReturnBot1.baseClasses.MessageConstants;
 import io.proj3ct.ReturnBot1.datebase.DatabaseConnection;
 
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
 /**
  * Класс, реализующий логику для удаления факультетов.
  * Он включает в себя обработку команд для удаления факультетов,
@@ -24,7 +29,7 @@ public class LogicForDeleteDepartment {
     private DatabaseConnection databaseConnection = new DatabaseConnection();
     /** Объект для получения данных из переменных окружения. */
     private EnvironmentService environmentService =new EnvironmentService();
-
+    private  Map<Long, String> messageTextForUser = new HashMap<>();
     public LogicForDeleteDepartment(DatabaseConnection databaseConnection,DataForDepartment dataForDepartment,
                                     DepartmentsInfo departmentsInfo,
                                     KeyboardsData keyboardsData, EnvironmentService environmentService) {
@@ -76,6 +81,29 @@ public class LogicForDeleteDepartment {
         }
         return messageTextForUser;
     }
+    private boolean checkValidNumberForDelete(String messageTextForUser, String messageText){
+        List<Integer> numbers = new ArrayList<>();
+
+        // Регулярное выражение для извлечения чисел
+        Pattern pattern = Pattern.compile("\\d+");
+        Matcher matcher = pattern.matcher(messageTextForUser);
+
+        // Найдем все числа и добавим их в список
+        while (matcher.find()) {
+            numbers.add(Integer.parseInt(matcher.group()));
+        }
+
+        try {
+            int messageTextInt = Integer.parseInt(messageText);
+            if(numbers.contains(messageTextInt)){
+                return false;
+            }
+        } catch (NumberFormatException e) {
+            return true;
+        }
+        return true;
+        //
+    }
     /**
      * Обрабатывает команды пользователя для удаления факультета.
      *
@@ -89,11 +117,15 @@ public class LogicForDeleteDepartment {
             userStatesForDeleteDepartment.put(userId, "awaiting_password");
             return MessageConstants.PASSWORD_COMMAND_RESPONSE;
         } else if ("awaiting_password".equals(currentState)) {
-            return checkValidPasswordInput(messageText, userId);
+            messageTextForUser.put(userId,checkValidPasswordInput(messageText, userId));
+            return messageTextForUser.get(userId);
         }
         else if("awaiting_numberForDelete".equals(currentState)){
-            dataForDepartment.setNumberForDeleteDepartment(userId, messageText);
             userStatesForDeleteDepartment.remove(userId);
+            if(checkValidNumberForDelete(messageTextForUser.get(userId), messageText)){
+                return MessageConstants.UN_CORRECT_NUMBER_COMMAND_RESPONSE;
+            }
+            dataForDepartment.setNumberForDeleteDepartment(userId, messageText);
             keyboardsData.deleteData(userId,databaseConnection,dataForDepartment);
             return MessageConstants.SUCCESSFUL_DELETE_DEPARTMENT_COMMAND_RESPONSE;
         }
